Series Details: reorganize columns hiding logic (#415)
For optional columns in the series details page, the code used to check
only base issues. This change makes colums appear if a value is set for
any issue (base or variant) and be hidden if the field is set to "no"
or empty for all issues in the series.