Get elasticsearch working for searches (#441)
* Downgrade elasticsearch back to 1.x

We'll need to do much more work to update past that, as the
libraries and interfaces have changed a lot, and django-haystack
is no longer maintained.

* No longer need to encode query for haystack

* Provide a sensible dev default for haystack

This immediately updates the search index, which is fine
for local installations.  We already set it to something else
in production (and presumably beta).